Quiet compiler warnings.

+++ AnimationControllers/CECardsAnimationController.m
@@ -42,7 +42,7 @@
         // push the from- view to the back
         [UIView addKeyframeWithRelativeStartTime:0.0f relativeDuration:0.4f animations:^{
             fromView.layer.transform = t1;
-            fromView.alpha = 0.6;
+            fromView.alpha = 0.6f;
         }];
         [UIView addKeyframeWithRelativeStartTime:0.2f relativeDuration:0.4f animations:^{
             fromView.layer.transform = t2;
@@ -73,8 +73,8 @@
     CGRect frame = [transitionContext initialFrameForViewController:fromVC];
     toView.frame = frame;
     CATransform3D scale = CATransform3DIdentity;
-    toView.layer.transform = CATransform3DScale(scale, 0.6, 0.6, 1);
-    toView.alpha = 0.6;
+    toView.layer.transform = CATransform3DScale(scale, 0.6f, 0.6f, 1.0f);
+    toView.alpha = 0.6f;
     
     [containerView insertSubview:toView belowSubview:fromView];
     
@@ -109,9 +109,9 @@
 
 -(CATransform3D)firstTransform{
     CATransform3D t1 = CATransform3DIdentity;
-    t1.m34 = 1.0/-900;
-    t1 = CATransform3DScale(t1, 0.95, 0.95, 1);
-    t1 = CATransform3DRotate(t1, 15.0f * M_PI/180.0f, 1, 0, 0);
+    t1.m34 = 1.0f/-900.0f;
+    t1 = CATransform3DScale(t1, 0.95f, 0.95f, 1.0f);
+    t1 = CATransform3DRotate(t1, 15.0f * (float)M_PI/180.0f, 1.0f, 0.0f, 0.0f);
     return t1;
     
 }
@@ -120,8 +120,8 @@
     
     CATransform3D t2 = CATransform3DIdentity;
     t2.m34 = [self firstTransform].m34;
-    t2 = CATransform3DTranslate(t2, 0, view.frame.size.height*-0.08, 0);
-    t2 = CATransform3DScale(t2, 0.8, 0.8, 1);
+    t2 = CATransform3DTranslate(t2, 0, view.frame.size.height*-0.08f, 0);
+    t2 = CATransform3DScale(t2, 0.8f, 0.8f, 1);
     
     return t2;
 }
